// graphql-program-apis/src/types.d.ts
export type ProposalType = {
  abstainVoteWeight?: number;
  closedAt?: number;
  denyVoteWeight?: number;
  descriptionLink?: string;
  draftAt?: number;
  executingAt?: number;
  executionFlags?: number;
  governance?: string;
  governingTokenMint?: string;
  lamports?: number;
  maxVoteWeight?: number;
  maxVotingTime?: number;
  name?: string;
  options?: Record<string, any>;
  pubkey: string;
  reserved1?: number;
  signatoriesCount?: number;
  signatoriesSignedOffCount?: number;
  signingOffAt?: number;
  startVotingAt?: number;
  state?: number;
  tokenOwnerRecord?: string;
  vetoVoteWeight?: number;
  voteThreshold?: Record<string, any>;
  voteType?: Record<string, any>;
  votingAt?: number;
  votingAtSlot?: number;
  votingCompletedAt?: number;
};